move base menus, constante_tributaria.py and sped_participante object, views
