+++ l10n_br_base/models/inherited_res_company.py
